snappy: move SnapFile.Install() into Overlord.Install()
This commit just moves the SnapFile.Install() from one place to
the next. Eventually this will remove "Install" from the part
interface.
