+++ snappy/snap_remote.go
@@ -202,12 +202,12 @@ func (s *RemoteSnapPart) Install(pbar progress.Meter, flags InstallFlags) (strin
 		return "", err
 	}
 
-	unauthOk := (flags & AllowUnauthenticated) != 0
-	snapFile, err := NewSnapFile(downloadedSnap, s.Origin(), unauthOk)
+	sf, err := (&Overlord{}).Install(downloadedSnap, s.Origin(), flags, pbar)
 	if err != nil {
 		return "", err
 	}
-	return snapFile.Install(pbar, flags)
+
+	return sf.Name(), nil
 }
 
 // SetActive sets the snap active
